Split up AuthenticationResolver into smaller resolvers around authentication subjects
